move header files into fs/axfs directory
There is no reason to keep the header files under the
include/linux directory because there are no APIs being
exported. Also, all other file system drivers (except cramfs)
keep all their header files with the source files.
